+++ src/exppp/pretty_alg.h
@@ -4,7 +4,7 @@
 #include <express/linklist.h>
 #include <express/scope.h>
 
-void ALGargs_out(Linked_List args, int level);
-void ALGscope_out(Scope s, int level);
+void ALGargs_out(Linked_List args, size_t level);
+void ALGscope_out(Scope s, size_t level);
 
 #endif /* PRETTY_ALG_H */
