+++ src/exppp/pretty_ref.c
@@ -8,7 +8,7 @@
 #include "pp.h"
 #include "pretty_ref.h"
 
-void REFout(Dictionary refdict, Linked_List reflist, char *type, int level)
+void REFout(Dictionary refdict, Linked_List reflist, char *type, size_t level)
 {
     Dictionary dict;
     DictionaryEntry de;
