+++ src/exppp/pretty_stmt.h
@@ -5,8 +5,8 @@
 #include <express/linklist.h>
 
 char *STMTto_string(Statement s);
-void STMT_out(Statement s, int level);
-void STMTlist_out(Linked_List stmts, int level);
+void STMT_out(Statement s, size_t level);
+void STMTlist_out(Linked_List stmts, size_t level);
 void STMTout(Statement s);
 int STMTto_buffer(Statement s, char *buffer, int length);
