+++ src/exppp/pretty_where.h
@@ -3,7 +3,7 @@
 
 #include <express/linklist.h>
 
-void WHERE_out(Linked_List wheres, int level);
+void WHERE_out(Linked_List wheres, size_t level);
 void WHEREout(Linked_List w);
 int WHEREto_buffer(Linked_List w, char *buffer, int length);
 char *WHEREto_string(Linked_List w);
